+++ tests/EnumCastTest.php
@@ -10,7 +10,7 @@ class EnumCastTest extends ApplicationTestCase
 {
     public function test_can_set_model_value_using_enum_instance()
     {
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = UserType::Moderator();
 
         $this->assertEquals(UserType::Moderator(), $model->user_type);
@@ -18,7 +18,7 @@ class EnumCastTest extends ApplicationTestCase
 
     public function test_can_set_model_value_using_enum_value()
     {
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = UserType::Moderator;
 
         $this->assertEquals(UserType::Moderator(), $model->user_type);
@@ -28,13 +28,13 @@ class EnumCastTest extends ApplicationTestCase
     {
         $this->expectException(InvalidEnumMemberException::class);
 
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = 5;
     }
 
     public function test_getting_model_value_returns_enum_instance()
     {
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = UserType::Moderator;
 
         $this->assertInstanceOf(UserType::class, $model->user_type);
@@ -42,7 +42,7 @@ class EnumCastTest extends ApplicationTestCase
 
     public function test_can_get_and_set_null_on_enum_castable()
     {
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = null;
 
         $this->assertNull($model->user_type);
@@ -50,7 +50,7 @@ class EnumCastTest extends ApplicationTestCase
 
     public function test_that_model_with_enum_can_be_cast_to_array()
     {
-        $model = app(Example::class);
+        $model = new Example;
         $model->user_type = UserType::Moderator();
 
         $this->assertSame(['user_type' => 1], $model->toArray());