+++ tests/Models/Example.php
@@ -5,6 +5,9 @@ namespace BenSampo\Enum\Tests\Models;
 use BenSampo\Enum\Tests\Enums\UserType;
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @property UserType $user_type
+ */
 class Example extends Model
 {
     public $timestamps = false;