+++ tests/Models/NativeCastModel.php
@@ -6,6 +6,10 @@ use BenSampo\Enum\Tests\Enums\UserType;
 use BenSampo\Enum\Tests\Enums\UserTypeCustomCast;
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @property UserType $user_type
+ * @property UserTypeCustomCast $user_type_custom
+ */
 class NativeCastModel extends Model
 {
     protected $casts = [